figure refactor seems to work
