+++ examples/desktop/gridplot/gridplot.py
@@ -10,25 +10,23 @@ import fastplotlib as fpl
 import imageio.v3 as iio
 
 
-plot = fpl.GridPlot(shape=(2, 2))
-# to force a specific framework such as glfw:
-# plot = fpl.GridPlot(canvas="glfw")
+fig = fpl.Figure(shape=(2, 2))
 
 im = iio.imread("imageio:clock.png")
 im2 = iio.imread("imageio:astronaut.png")
 im3 = iio.imread("imageio:coffee.png")
 im4 = iio.imread("imageio:hubble_deep_field.png")
 
-plot[0, 0].add_image(data=im)
-plot[0, 1].add_image(data=im2)
-plot[1, 0].add_image(data=im3)
-plot[1, 1].add_image(data=im4)
+fig[0, 0].add_image(data=im)
+fig[0, 1].add_image(data=im2)
+fig[1, 0].add_image(data=im3)
+fig[1, 1].add_image(data=im4)
 
-plot.show()
+fig.show()
 
-plot.canvas.set_logical_size(800, 800)
+fig.canvas.set_logical_size(800, 800)
 
-for subplot in plot:
+for subplot in fig:
     subplot.auto_scale()
 
 if __name__ == "__main__":
