+++ examples/desktop/line/line.py
@@ -10,9 +10,7 @@ import fastplotlib as fpl
 import numpy as np
 
 
-plot = fpl.Plot()
-# to force a specific framework such as glfw:
-# plot = fpl.Plot(canvas="glfw")
+fig = fpl.Figure()
 
 xs = np.linspace(-10, 10, 100)
 # sine wave
@@ -28,20 +26,20 @@ a = 0.5
 ys = np.sinc(xs) * 3 + 8
 sinc = np.dstack([xs, ys])[0]
 
-sine_graphic = plot.add_line(data=sine, thickness=5, colors="magenta")
+sine_graphic = fig[0, 0].add_line(data=sine, thickness=5, colors="magenta")
 
 # you can also use colormaps for lines!
-cosine_graphic = plot.add_line(data=cosine, thickness=12, cmap="autumn")
+cosine_graphic = fig[0, 0].add_line(data=cosine, thickness=12, cmap="autumn")
 
 # or a list of colors for each datapoint
 colors = ["r"] * 25 + ["purple"] * 25 + ["y"] * 25 + ["b"] * 25
-sinc_graphic = plot.add_line(data=sinc, thickness=5, colors=colors)
+sinc_graphic = fig[0, 0].add_line(data=sinc, thickness=5, colors=colors)
 
-plot.show()
+fig.show()
 
-plot.canvas.set_logical_size(800, 800)
+fig.canvas.set_logical_size(800, 800)
 
-plot.auto_scale()
+fig[0, 0].auto_scale()
 
 
 if __name__ == "__main__":
