+++ fastplotlib/__init__.py
@@ -1,12 +1,12 @@
 from pathlib import Path
 
-from .layouts import Plot, GridPlot
+from .utils.gui import run
 from .graphics import *
 from .graphics.selectors import *
 from .legends import *
-from .widgets import ImageWidget
+from .layouts import Figure
+# from .widgets import ImageWidget
 from .utils import config
-from .utils.gui import run
 
 import wgpu
 
@@ -21,9 +21,8 @@ if len(adapters) < 1:
 
 
 __all__ = [
-    "Plot",
-    "GridPlot",
+    "Figure",
     "run",
-    "ImageWidget",
+    # "ImageWidget",
     "Legend",
 ]
