--- fastplotlib/layouts/_frame/__init__.py
@@ -1,1 +1,0 @@
-from ._frame import Frame
