+++ fastplotlib/layouts/output/_qtoolbar_template.py
