+++ fastplotlib/layouts/output/_toolbar.py
@@ -2,8 +2,8 @@ from .._subplot import Subplot
 
 
 class ToolBar:
-    def __init__(self, plot):
-        self.plot = plot
+    def __init__(self, figure):
+        self.figure = figure
 
     def _get_subplot_dropdown_value(self) -> str:
         raise NotImplemented
@@ -11,17 +11,17 @@ class ToolBar:
     @property
     def current_subplot(self) -> Subplot:
         """Returns current subplot"""
-        if hasattr(self.plot, "_subplots"):
+        if hasattr(self.figure, "_subplots"):
             # parses dropdown or label value as plot name or position
             current = self._get_subplot_dropdown_value()
             if current[0] == "(":
                 # str representation of int tuple to tuple of int
                 current = tuple(int(i) for i in current.strip("()").split(","))
-                return self.plot[current]
+                return self.figure[current]
             else:
-                return self.plot[current]
+                return self.figure[current]
         else:
-            return self.plot
+            return self.figure
 
     def panzoom_handler(self, ev):
         raise NotImplemented
