+++ fastplotlib/layouts/output/qtoolbar.ui
